⚒ Add connection point and path configuration in LayoutEngine (#7)
In the PR #6, I implemented edge drawing (but only for simple cases). At
that time, all processing was implemented in `Backend`, but this means
that the same processing needs to be reimplemented each time a new
backend is added. I want to move the following processes to
`LayoutEngine`.

- Placement of connection points
- Construction of paths when connecting appropriate connection points to
each other

I am also inclined to separate the placement of nodes from the
construction of connection paths. However, I don't want to make it more
complicated than necessary, so I will add another method to the layout
engine to handle that case.